fix: Breaking changes after the split
