+++ app/src/main/java/eu/kanade/tachiyomi/util/chapter/ChapterRecognition.kt
@@ -2,6 +2,7 @@ package eu.kanade.tachiyomi.util.chapter
 
 import eu.kanade.tachiyomi.source.model.SChapter
 import eu.kanade.tachiyomi.source.model.SManga
+import eu.kanade.tachiyomi.source.models.originalTitle
 
 /**
  * -R> = regex conversion.
